requirements.txt
openai-whisper
pydub
fuzzywuzzy
python-Levenshtein
ffmpeg-python
